requirements.txt
yfinance
pandas
numpy
matplotlib
xgboost
lightgbm
scikit-learn
discord.py
pillow
pytickertick
vaderSentiment
prophet
